Break amp-list routes out into own routing file
